refactor: 댓글 서비스 분리, 베스트 댓글 로직 수정 - #7
