+++ src/main/java/kr/inuappcenterportal/inuportal/domain/reply/repository/ReplyRepository.java
@@ -4,6 +4,8 @@ import kr.inuappcenterportal.inuportal.domain.member.model.Member;
 import kr.inuappcenterportal.inuportal.domain.post.model.Post;
 import kr.inuappcenterportal.inuportal.domain.reply.model.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,5 +21,14 @@ public interface ReplyRepository extends JpaRepository<Reply,Long> {
     Optional<Reply> findByMember(Member member);
 
     Optional<Reply> findFirstByMember(Member member);
+
+    // 하나의 게시물의 댓글 들 중에 좋아요가 많이 박힌 댓글 순으로 정렬
+    @Query("SELECT r FROM Reply r " +
+            "JOIN ReplyLike rl ON rl.reply = r " +
+            "WHERE r.post.id = :postId " + // 특정 Post에 속한 댓글 필터링
+            "GROUP BY r " +
+            "HAVING COUNT(rl) > 0 " +      // 좋아요가 1개 이상인 댓글만
+            "ORDER BY COUNT(rl) DESC")    // 좋아요 수 기준 내림차순 정렬
+    List<Reply> findRepliesWithLikesByPost(@Param("postId") Long postId);
 
 }
