+++ src/main/java/kr/inuappcenterportal/inuportal/global/config/SecurityConfig.java
@@ -36,7 +36,7 @@ public class SecurityConfig {
                 .authorizeHttpRequests(auth->auth.requestMatchers("/swagger", "/swagger-ui.html", "/swagger-ui/**", "/api-docs", "/api-docs/**", "/v3/api-docs/**").permitAll()
                         .requestMatchers(HttpMethod.POST,"/api/members/**","/api/members").permitAll()
                         .requestMatchers(HttpMethod.POST,"/api/reports/**").hasAnyRole("USER","ADMIN")
-                        .requestMatchers("/api/members/all","/api/reports").hasRole("ADMIN")
+                        .requestMatchers("/api/members/all","/api/reports").hasRole("USER") // 테스트 용도로 ADMIN -> USER
                         .requestMatchers("/api/members/**","/api/members").hasAnyRole("USER","ADMIN")
                         .requestMatchers(HttpMethod.GET,"/api/posts/**","/api/posts","/api/cafeterias","/api/weathers").permitAll()
                         .requestMatchers("/api/posts/**","/api/posts","/api/fires/**").hasAnyRole("USER","ADMIN")
@@ -44,7 +44,7 @@ public class SecurityConfig {
                         .requestMatchers("/api/search","/api/notices","/api/notices/**","api/schedules","api/schedules/**").permitAll()
                         .requestMatchers("/api/folders/**","/api/folders","/api/search/**").hasAnyRole("USER","ADMIN")
                         .requestMatchers(HttpMethod.GET,"/api/categories","/api/images/**").permitAll()
-                        .requestMatchers("/api/images","/api/images/**","/api/categories").hasRole("ADMIN"));
+                        .requestMatchers("/api/images","/api/images/**","/api/categories").hasRole("USER")); // 테스트 용도로 ADMIN -> USER
         httpSecurity
                 .addFilterBefore(new JwtAuthenticationFilter(tokenProvider,objectMapper), UsernamePasswordAuthenticationFilter.class)
                 .exceptionHandling(exception->exception.accessDeniedHandler(new CustomAccessDeniedHandler(objectMapper))
