+++ src/main/java/kr/inuappcenterportal/inuportal/global/service/RedisService.java
@@ -111,9 +111,6 @@ public class RedisService {
         }
     }
 
-
-
-
     public void storeMeal(String cafeteria,int day, int num,String menu){
         String key = cafeteria+"-"+day+"-"+num;
         redisTemplate.opsForValue().set(key,menu);
