+++ src/test/java/kr/inuappcenterportal/inuportal/controller/MemberControllerTest.java
@@ -14,7 +14,7 @@ import kr.inuappcenterportal.inuportal.global.exception.ex.MyErrorCode;
 import kr.inuappcenterportal.inuportal.global.exception.ex.MyException;
 import kr.inuappcenterportal.inuportal.domain.member.service.MemberService;
 import kr.inuappcenterportal.inuportal.domain.post.service.PostService;
-import kr.inuappcenterportal.inuportal.domain.reply.service.ReplyService;
+import kr.inuappcenterportal.inuportal.domain.reply.service.ReplyCommandService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -57,7 +57,7 @@ public class MemberControllerTest {
     @MockBean
     MemberService memberService;
     @MockBean
-    ReplyService replyService;
+    ReplyCommandService replyService;
     @MockBean
     PostService postService;
